LayoutTests/gamepad/gamepad-api.html

<!DOCTYPE html>
<body>
<script src="gamepad-test.js"></script>
<script>
    testExpected("navigator.webkitGamepads", undefined, "!==");
    endTest();
</script>
<p>Make sure the main polling access point exists on navigator.</p>
</body>
