LayoutTests/gamepad/gamepad-api-expected.txt

EXPECTED (navigator.webkitGamepads !== 'undefined') OK
END OF TEST
Make sure the main polling access point exists on navigator.
